+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Added
 
+- BC-252 - creating componenets and grid for rooms-overview page
 - BC-98 - added unit tests for school admin page
 
 ### Changed